Include apex_pubkey in built APEX
This is never read for signature verification, but I didn't realize it
does get read when determining which staged APEXes can be activated.
Without it, the installation appears to succeed but the APEX disappears
without a trace.

Protip: find apexd logs in early dmesg. You might need to increase the
buffer size with the kernel's log_buf_len=,